+++ trane/datasets/__init__.py
@@ -1,8 +1,6 @@
 from trane.datasets.load_functions import (
     load_covid,
     load_covid_metadata,
-    load_bike,
-    load_bike_metadata,
     load_youtube,
     load_youtube_metadata,
 )