+++ trane/ops/op_base.py
@@ -2,8 +2,6 @@ import pandas as pd
 
 from trane.ops.threshold_functions import sample_unique_values
 
-__all__ = ["OpBase"]
-
 
 class OpBase(object):
     """